Moved List component out of main folder
